requirements.txt
pyyaml>=6.0.1
jsonschema>=4.22.0
svgwrite>=1.4.3
cairosvg>=2.7.0
Pillow>=10.3.0
numpy>=1.26.4